requirements.txt
streamlit
opencv-python-headless
easyocr
fpdf2
numpy
Pillow